Move watch info classes to their own subpackage in watchmanager
